# openapi.yaml
openapi: "3.0.3"
info:
  title: ColourYou API
  description: "L'API du moment"
  version: 2.0.0
# servers:
#   - url: http://api.example.com/v1
#     description: Optional server description, e.g. Main (production) server
#   - url: http://staging-api.example.com
#     description: Optional server description, e.g. Internal staging server for testing
...
paths:
...
  /couleurs:
    get:
      tags:
        - couleur
      summary: Retourne une liste des couleurs
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nom de la couleur
                  example: Bleu marine
                hex:
                  type: string
                  description: Code hexadécimal de la couleur
                  example: '#1F1F1F'
                datePosted:
                  type: date
                  description: Date à laquelle la couleur apparaît
              examples:
                name: Vert pomme
                hex: '#222222'
                datePosted: 07/09/2022
      responses:
        '200':
          description: OK
        '400':
          description: ''
        '404':
          description: Page non trouvée
    put:
      tags:
        - couleur
      summary: Met à jour les couleurs
      description: ''
  /couleurs/{id}:
    get:
      tags:
        - couleur
      summary: Affiche la couleur par son ID
      description: ''
      parameters:
        - name: photoId
          in: path
          description: ID de la couleur
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nom de la couleur
                  example: Bleu marine
                hex:
                  type: string
                  description: Code hexadécimal de la couleur
                  example: '#1F1F1F'
                datePosted:
                  type: date
                  description: 07/09/2022
    post:
      tags:
        - couleur
      summary: Modifier la couleur par son ID
      description: ''
      parameters:
        - name: couleurId
          in: path
          description: ID de la couleur
          required: true
          unique: true
          schema:
            type: string
        - name: name
          in: path
          description: Nom de la couleur
          required: true
          unique: true
          schema:
            type: string
        - name: hex
          in: path
          description: Code hexadécimal de la couleur
          required: true
          unique: true
          schema:
            type: string
        - name: datePosted
          in: path
          description: Date à laquelle la couleur apparaît
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Nom de la couleur
                  example: Bleu marine
                hex:
                  type: string
                  description: Code hexadécimal de la couleur
                  example: '#1F1F1F'
                datePosted:
                  type: date
                  description: Date à laquelle la couleur apparaît
      responses:
        '200':
          description: Couleur ajoutée avec succès
        '400':
          description: ''
        '404':
          description: Page non trouvée
    delete:
      tags:
        - couleur
      summary: Supprimer la couleur par son ID
      description: ''